refactor : 피드백 반영  [AIR-112]
- BasicControllerTest -> BaseCOntrollerTest로 통합
- JwtAuthenticationFilter 오타 수정
- Tokens Util성 메서드였던 getToken을 resolver의 private 메서드로 변경
